Delete SuBSENSE algorithm (motion saliency) due to license issues
Part of SuBSENSE algorithm is based on ViBe method, that is patented in
Europe ans U.S.
This means that SuBSENSE can not currently be released under the BSD
license. We are trying to resolve the issue.
In waiting, I am forced to remove the algorithm from the "official"
branch on which the pull-request on opencv_contrib will be made . The
code has been moved to an opportune branch (that will not be merged) and
will remain there waiting for this issue to be clarified.