matlab: fix binding generation for modules that are part of opencv_contrib (xfeatures2d)
This patch fixes Matlab binding generation for modules that are
part of opencv_contrib (xfeatures2d specifically), and are therefore
not located in ${CMAKE_SOURCE_DIR}/modules.

For this, gen_matlab.py script is extended to accept multiple values
for --modules-root argument, and to check for a module's header in
all possible paths. This way, both ${CMAKE_SOURCE_DIR}/modules and
${OPENCV_EXTRA_MODULES_PATH} can be passed as modules root paths,
and main header for xfeatures2d module is correctly located.