Enabled FAST feature detection on specific image locations
Some applications only want to know if there are feature at specific
locations. To fit these needs the FAST approach is extended in
FASTForPointSet to recieve a vector of locations and calculates the FAST
response on these positions. 
If it is below the threshold, it will be removed from the list.