This commit contains all significant contributions until 23/8/2016
Implemented ImagePreprocessor a geneology of functors for preprocessing
samples before they are fed to a CNN.

Implemented TextImageClassifer as an abstraction of any object that
provides a vector of probabillities given an image. A gluing class
providing compatibillity to callbacks for OCRBeamsearchDecoder allows
to plug any such object to the existing algorithms.

Implemented DeepCNN a class that is supposed to load pretrained caffe
models. Although several backends have been planed the only fully ope-
rational one is the one employing caffe its self. GPU is supported.

Implemented a class using a DeepCNN that implements word-spotting

Implemented a python and a C++ demo demonstraiting the WorSpotting
Class

Implemented a class that generates sythetic text samples at high speeds
and with a full pipeline of realistic distortions for training
Word-spotting CNN's who need millios of samples.

Implemented a python demo demonstraiting the usage of the text synthesizer