Comment out direct print statements during tesseract run invocations
This statement interferes with regular logging on the client side
where the OpenCV library's outputs cannot be managed by the user
and forcefully print out in the client code. For instance, this
is observed when OpenCV is used as a python library where it is
expected to provide optional loggers instead of direct writes to
stdout.