optflow(ocl): avoid rebuilding kernels
Rebuilding the same static kernel every time a function is called is not
exactly idiomatic in OpenCL, and unless the CL implementation has a fast
cache setup for kernels it introduces a lot of overhead. This moves
building of the TVL1 optical flow kernels into the tvl1flow object's
constructor, allowing them to be reused within that object.