feat: Modernized python code using `futurize`
[futurize](https://www.python-future.org/futurize.html) is a tool that is designed
to make python 2 code compatible with both python 2 and python 3.  This commit contains
the output of that tool, which I've inspected and manually adjusted to remove
spurious fixes (e.g., print statements that were written as functions in the python 2
code would get an extra set of parenthesis which needed to be manually removed).